+++ tests/Jest/components/TestRail/Result.spec.js
@@ -20,3 +20,17 @@ test('Elapsed time is correctly converted from MS to S', () => {
     const result = new Result('C123', 1, 'this is a test', 15000);
     expect(result.getElapsed()).toBe('15s');
 });
+
+test('Screenshots are empty if not assigned', () => {
+    const result = new Result('C123', 1, 'this is a test', 15000);
+    expect(result.getScreenshotPaths().length).toBe(0);
+});
+
+test('Screenshots are correctly assigned', () => {
+    const result = new Result('C123', 1, 'this is a test', 15000, ['path1', 'path2']);
+
+    expect(result.getScreenshotPaths().length).toBe(2);
+    expect(result.getScreenshotPaths()[0]).toBe('path1');
+    expect(result.getScreenshotPaths()[1]).toBe('path2');
+});
+